requirements.txt
Django>=3.2.25,<3.3
djangorestframework>=3.14.0,<3.15
drf-yasg>=1.21.7,<1.22
mysqlclient>=2.2.4,<2.3
mysql-connector-python>=8.4.0,<8.5
redis>=5.0.4,<5.1
celery>=5.4.0,<5.5
django-celery-results>=2.5.1,<2.6
python-telegram-bot>=21.2,<21.3
beautifulsoup4>=4.12.3,<4.13
selenium>=4.21.0,<4.22
undetected-chromedriver>=3.5.5,<3.6
asgiref>=3.8.1,<3.9
django-adminlte-3>=0.1.6,<0.2